Refactor LoadingSpinner component and update import paths for consistency
