+++ src/app/print-preview/bill/both/page.tsx
@@ -1,6 +1,6 @@
 'use client';
+import LoadingSpinner from '@/app/components/LoadingSpinner';
 import CustomerBillPreview from '@/app/print-preview/components/CustomerBillPreview';
-import LoadingSpinner from '@/app/print-preview/components/LoadingSpinner';
 import PrintHeader from '@/app/print-preview/components/PrintHeader';
 import WorkerBillPreview from '@/app/print-preview/components/WorkerBillPreview';
 import handleError from '@/app/util/error/handleError';
@@ -49,7 +49,7 @@ const BothBillPage: React.FC = () => {
   return (
     <>
       {!isDataLoaded ? (
-        <LoadingSpinner /> // Use the LoadingSpinner component
+        <LoadingSpinner classStyle="h-screen" />
       ) : (
         <>
           <PrintHeader backUrl={backUrl} isLoading={!isDataLoaded} />
